ISSUE #4780 - rename components/folder to include to specify it's container revision
