+++ frontend/src/v5/services/api/index.ts
@@ -22,9 +22,9 @@ import * as Teamspaces from './teamspaces';
 import * as Projects from './projects';
 import * as Containers from './containers';
 import * as Federations from './federations';
-import * as Revisions from './revisions';
+import * as ContainerRevisions from './containerRevisions';
 import * as Users from './users';
 import * as Tickets from './tickets';
 import * as Drawings from './drawings';
 
-export { Auth, TicketComments, CurrentUser, Teamspaces, Projects, Containers, Federations, Revisions, Users, Tickets, Drawings };
+export { Auth, TicketComments, CurrentUser, Teamspaces, Projects, Containers, Federations, ContainerRevisions as Revisions, Users, Tickets, Drawings };
