+++ frontend/src/v5/services/routing/routing.tsx
@@ -18,7 +18,7 @@ import { IContainer } from '@/v5/store/containers/containers.types';
 import { IFederation } from '@/v5/store/federations/federations.types';
 import { IProject } from '@/v5/store/projects/projects.types';
 import { generatePath } from 'react-router';
-import { IRevision } from '@/v5/store/revisions/revisions.types';
+import { IContainerRevision } from '@/v5/store/containerRevisions/containerRevisions.types';
 import { VIEWER_ROUTE, PROJECT_ROUTE_BASE, PROJECT_ROUTE, BOARD_ROUTE } from '@/v5/ui/routes/routes.constants';
 
 export const projectRoute = (teamspace: string, project: IProject | string) => {
@@ -31,7 +31,7 @@ export const projectTabRoute = (teamspace: string, project: IProject | string, t
 	return generatePath(PROJECT_ROUTE, { teamspace, project: projectId, tab });
 };
 
-type RevisionParam = IRevision | string | null | undefined;
+type RevisionParam = IContainerRevision | string | null | undefined;
 type ContainerOrFederationParam = IContainer | IFederation | string;
 
 export const viewerRoute = (
@@ -43,7 +43,7 @@ export const viewerRoute = (
 	const containerOrFederationId = (containerOrFederation as IContainer | IFederation)?._id
 		|| (containerOrFederation as string);
 
-	const revisionId = (revision as IRevision)?._id || (revision as string);
+	const revisionId = (revision as IContainerRevision)?._id || (revision as string);
 
 	const params = {
 		teamspace,
