+++ frontend/src/v5/services/selectorsHooks.ts
@@ -25,8 +25,8 @@ import * as ProjectsSelectors from '@/v5/store/projects/projects.selectors';
 import * as FederationsSelectors from '@/v5/store/federations/federations.selectors';
 import * as GroupsSelectors from '@/v4/modules/groups/groups.selectors';
 import * as ModelSelectors from '@/v4/modules/model/model.selectors';
-import * as RevisionsSelectors from '@/v5/store/revisions/revisions.selectors';
 import * as RouterSelectors from '@/v4/modules/router/router.selectors';
+import * as ContainerRevisionsSelectors from '@/v5/store/containerRevisions/containerRevisions.selectors';
 import * as SequencesSelctors from '@/v4/modules/sequences/sequences.selectors';
 import * as TreeSelectors from '@/v4/modules/tree/tree.selectors';
 import * as TicketsSelectors from '@/v5/store/tickets/tickets.selectors';
@@ -49,8 +49,8 @@ export const FederationsHooksSelectors = createHooksSelectors(FederationsSelecto
 export const GroupsHooksSelectors = createHooksSelectors(GroupsSelectors);
 export const ModelHooksSelectors = createHooksSelectors(ModelSelectors);
 export const ProjectsHooksSelectors = createHooksSelectors(ProjectsSelectors);
-export const RevisionsHooksSelectors = createHooksSelectors(RevisionsSelectors);
 export const RouterHooksSelectors = createHooksSelectors(RouterSelectors);
+export const ContainerRevisionsHooksSelectors = createHooksSelectors(ContainerRevisionsSelectors);
 export const SequencesHooksSelectors = createHooksSelectors(SequencesSelctors);
 export const TicketsHooksSelectors = createHooksSelectors(TicketsSelectors);
 export const TicketsCardHooksSelectors = createHooksSelectors(TicketsCardSelectors);
