--- frontend/src/v5/ui/components/shared/revisionDetails/components/skeletonListItem/skeletonListItem.styles.ts
@@ -1,34 +1,0 @@
-/**
- *  Copyright (C) 2021 3D Repo Ltd
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-import styled from 'styled-components';
-
-import { DashboardListSkeletonItem } from '@components/dashboard/dashboardList/dashboardListSkeletonItem';
-import { SkeletonBlock as SkeletonBlockBase } from '@controls/skeletonBlock/skeletonBlock.styles';
-
-export const Container = styled(DashboardListSkeletonItem)`
-	height: 48px;
-	background-color: ${({ theme }) => theme.palette.secondary.light};
-	padding-left: 20px;
-	margin-left: 66px;
-	margin-bottom: 1px;
-	width: auto;
-`;
-
-export const SkeletonBlock = styled(SkeletonBlockBase)`
-	height: 8px;
-	background-color: ${({ theme }) => theme.palette.secondary.lightest};
-`;
