+++ frontend/src/v5/ui/routes/dashboard/projects/containers/containersList/containerListItem/containerEllipsisMenu/containerEllipsisMenu.component.tsx
@@ -26,7 +26,7 @@ import { DashboardParams } from '@/v5/ui/routes/routes.constants';
 import { ContainersHooksSelectors, ProjectsHooksSelectors } from '@/v5/services/selectorsHooks';
 import { prefixBaseDomain } from '@/v5/helpers/url.helper';
 import { ContainerSettingsModal } from '../../../containerSettingsModal/containerSettingsModal.component';
-import { uploadToContainer } from '../../../uploadFileForm/uploadFileForm.helpers';
+import { uploadToContainer } from '../../../uploadContainerRevisionForm/uploadContainerRevisionForm.helpers';
 
 type ContainerEllipsisMenuProps = {
 	selected?: boolean,
