--- frontend/src/v5/ui/routes/dashboard/projects/containers/uploadFileForm/uploadFileForm.helpers.ts
@@ -1,37 +1,0 @@
-/**
- *  Copyright (C) 2022 3D Repo Ltd
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU Affero General Public License as
- *  published by the Free Software Foundation, either version 3 of the
- *  License, or (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU Affero General Public License for more details.
- *
- *  You should have received a copy of the GNU Affero General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { DialogsActionsDispatchers } from '@/v5/services/actionsDispatchers';
-import { uploadFile, getSupportedFileExtensions } from '@controls/fileUploader/uploadFile';
-import { UploadFileForm } from './uploadFileForm.component';
-
-export const extensionIsSpm = (extension: string) => extension === 'spm';
-
-export const reduceFileData = (files) => files.map(({ file: { name, size }, ...rest }) => ({ file: { name, size }, ...rest }));
-
-export const extensionIsRevit = (extension: string) => ['rvt', 'rfa'].includes(extension);
-
-export const uploadToContainer = async (presetContainerId: string) => {
-	const onUpload = (presetFile) => {
-		DialogsActionsDispatchers.open(UploadFileForm, {
-			presetFile,
-			presetContainerId,
-		});
-	};
-	const file = await uploadFile(getSupportedFileExtensions());
-	onUpload(file);
-};
