+++ frontend/test/containers/containers.store.spec.ts
@@ -20,7 +20,7 @@ import { selectCanUploadToProject, selectContainerById, selectContainers, select
 import { times } from 'lodash';
 import { containerMockFactory, prepareMockSettingsReply, prepareMockStats, prepareMockViews } from './containers.fixtures';
 import { NewContainer, UploadStatuses } from '@/v5/store/containers/containers.types';
-import { revisionsMockFactory } from '../revisions/revisions.fixtures';
+import { containerRevisionsMockFactory } from '../containerRevisions/containerRevisions.fixtures';
 import { ProjectsActions } from '@/v5/store/projects/projects.redux';
 import { createTestStore, listContainsElementWithId } from '../test.helpers';
 import { Role } from '@/v5/store/currentUser/currentUser.types';
@@ -117,7 +117,7 @@ describe('Containers: store', () => {
 		});
 
 		it('should update revision processing status', () => {
-			const newRevision = revisionsMockFactory();
+			const newRevision = containerRevisionsMockFactory();
 			const newContainer = createAndAddContainerToStore();
 			dispatch(ContainersActions.containerProcessingSuccess(projectId, newContainer._id, newRevision));
 			const containerFromState = selectContainerById(getState(), newContainer._id);
